refactor: related to "Legacy: v2.0 Content Collections API"
- See: https://docs.astro.build/en/guides/upgrade-to/v5/#legacy-v20-content-collections-api
