+++ .vscode/astro-antfustyle-theme.code-snippets
@@ -46,10 +46,11 @@
     "prefix": "projectData",
     "body": [
       "{",
-      "  \"name\": \"${1:Name of the project to be displayed}\",",
+      "  \"id\": \"${1:Name of the project to be displayed}\",",
       "  \"link\": \"${2:URL linking to the project page or repository}\",",
       "  \"desc\": \"${3:A brief description summarizing the project}\",",
       "  \"icon\": \"${4:Icon representing the project. It must be in the format `i-<collection>-<icon>` or `i-<collection>:<icon>` as per [UnoCSS](https://unocss.dev/presets/icons) specs. [Check all available icons here](https://icones.js.org/)}\"",
+      "  \"category\": \"${5:Category of the project}\",",
       "}",
     ],
     "description": "Insert a project entry for projects data",
@@ -60,7 +61,7 @@
     "prefix": "streamData",
     "body": [
       "{",
-      "  \"title\": \"${1:Sets the stream title}\",",
+      "  \"id\": \"${1:Sets the stream title}\",",
       "  \"pubDate\": \"${2:$CURRENT_YEAR-$CURRENT_MONTH-$CURRENT_DATE}\",",
       "  \"link\": \"${3:Specifies the URL link to the stream}\",",
       "  \"radio\": ${4|false,true|},",
